# GREEN BUILD PROJECT/openapi.yaml
openapi: 3.0.3
info:
  title: EcoBuild API
  version: 1.0.0
  description: |-
    REST API for EcoBuild marketplace, including auth, products, image upload, metrics, and approval workflow.
servers:
  - url: http://localhost:5000
paths:
  /api/signup:
    post:
      summary: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, role]
              properties:
                role:
                  type: string
                  enum: [buyer, supplier, admin]
                email:
                  type: string
                  format: email
                password:
                  type: string
                full_name:
                  type: string
                phone:
                  type: string
                organization:
                  type: string
                company:
                  type: string
                location:
                  type: string
      responses:
        '200':
          description: Account created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user_id:
                    type: integer
  /api/login:
    post:
      summary: Authenticate and start a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, role]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                role:
                  type: string
                  enum: [buyer, supplier, admin]
      responses:
        '200':
          description: Authenticated user and CSRF token
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                  full_name:
                    type: string
                  role:
                    type: string
                  email:
                    type: string
                  csrf_token:
                    type: string
  /api/logout:
    post:
      summary: Log out current session
      security:
        - cookieAuth: []
      parameters:
        - name: X-CSRF-Token
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Logged out successfully
  /api/session:
    get:
      summary: Get current authenticated session info
      responses:
        '200':
          description: Current authenticated user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                  full_name:
                    type: string
                  role:
                    type: string
                  email:
                    type: string
                  csrf_token:
                    type: string
  /api/products:
    get:
      summary: List products with optional filters
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, price, carbon]
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      summary: Create a new product
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      parameters:
        - name: X-CSRF-Token
          in: header
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Product created successfully
  /api/products/{product_id}:
    put:
      summary: Update an existing product
      security:
        - cookieAuth: []
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
        - name: X-CSRF-Token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '200':
          description: Product updated successfully
    delete:
      summary: Delete a product
      security:
        - cookieAuth: []
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
        - name: X-CSRF-Token
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product deleted successfully
  /api/products/pending:
    get:
      summary: List pending products for admin review
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Pending products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /api/products/{product_id}/approve:
    post:
      summary: Approve a pending product
      security:
        - cookieAuth: []
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
        - name: X-CSRF-Token
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product approved
  /api/products/{product_id}/reject:
    post:
      summary: Reject a pending product
      security:
        - cookieAuth: []
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
        - name: X-CSRF-Token
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product rejected
  /api/upload-image:
    post:
      summary: Upload a product image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Uploaded image URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  image_url:
                    type: string
  /api/metrics:
    get:
      summary: Retrieve marketplace metrics
      responses:
        '200':
          description: Metrics results
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_products:
                    type: integer
                  average_carbon:
                    type: number
                  total_suppliers:
                    type: integer
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
        supplier_id:
          type: integer
          nullable: true
        name:
          type: string
        supplier:
          type: string
        type:
          type: string
        price:
          type: number
        strength:
          type: number
          nullable: true
        durability:
          type: number
          nullable: true
        carbon:
          type: number
        unit:
          type: string
        image_url:
          type: string
          nullable: true
        status:
          type: string
        created_at:
          type: string
    ProductInput:
      type: object
      required: [name, supplier, type, price, carbon, unit]
      properties:
        name:
          type: string
        supplier:
          type: string
        type:
          type: string
        price:
          type: number
        strength:
          type: number
          nullable: true
        durability:
          type: number
          nullable: true
        carbon:
          type: number
        unit:
          type: string
        image_url:
          type: string
          nullable: true